+++ requirements.txt
@@ -1,2 +1,7 @@
 azure-mgmt-resource>=18.0.0
+azure-mgmt-synapse>=2.0.0
+azure-synapse>=0.1.1
 azure-identity>=1.5.0
+pyspark
+delta-spark
+faker
